+++ BackEnd/ProyectoJava/PI-BackEnd/src/main/java/com/ctd/proyectointegrador/persistance/repository/ReservaRepository.java
@@ -4,5 +4,5 @@ package com.ctd.proyectointegrador.persistance.repository;
 import com.ctd.proyectointegrador.persistance.model.Reserva;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservaRepository extends JpaRepository<Reserva, Integer> {
+public interface ReservaRepository extends JpaRepository<Reserva, Long> {
 }
