+++ BackEnd/ProyectoJava/PI-BackEnd/src/main/java/com/ctd/proyectointegrador/service/IService.java
@@ -5,11 +5,11 @@ public interface IService<T> {
 
     public Map<String, Object> guardar(T object);
 
-    public Map<String, Object> buscar(Integer id);
+    public Map<String, Object> buscar(Long id);
 
-    public Map<String, Object> actualizar(Integer id, T object);
+    public Map<String, Object> actualizar(Long id, T object);
 
-    public Map<String, Object> eliminar(Integer id);
+    public Map<String, Object> eliminar(Long id);
 
     public Map<String, Object> listarTodos();
 }
